fix: (SPUI-571): fix comments from github
